add new var $enable_install_tool (default: false)
ATTENTION:
If var is set to true:
This adds the ENABLE_INSTALL_TOOL file in typo3conf, if set to true.
The file will not be changed, but will recreated it. The deletion causes a new
file is created.

If the var is set to false:
The ENABLE_INSTALL_TOOL file is deleted on next puppet run when it is older
than 1 hour. If a backend user creates the ENABLE_INSTALL_TOOL, he has enough
time to do his job.